Bring back data providers, minor fixes
