+++ tests/Acceptance/Service/UpdateByQueryRequestTest.php
@@ -10,18 +10,19 @@ use Answear\LuigisBoxBundle\Service\Client;
 use Answear\LuigisBoxBundle\Service\LuigisBoxSerializer;
 use Answear\LuigisBoxBundle\Service\UpdateByQueryRequest;
 use Answear\LuigisBoxBundle\Service\UpdateByQueryRequestInterface;
+use Answear\LuigisBoxBundle\Tests\DataProvider\UpdateByQueryRequestDataProvider;
 use Answear\LuigisBoxBundle\Tests\ExampleConfiguration;
 use Answear\LuigisBoxBundle\ValueObject\UpdateByQuery;
 use GuzzleHttp\Psr7\Response;
 use GuzzleHttp\Psr7\Uri;
-use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\DataProviderExternal;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 
 class UpdateByQueryRequestTest extends TestCase
 {
     #[Test]
-    #[DataProvider('forUpdate')]
+    #[DataProviderExternal(UpdateByQueryRequestDataProvider::class, 'forUpdate')]
     public function updatePassed(
         UpdateByQuery $updateByQuery,
         string $expectedContent,
@@ -36,35 +37,8 @@ class UpdateByQueryRequestTest extends TestCase
         $this->assertSame($jobId, $response->jobId);
     }
 
-    public static function forUpdate(): iterable
-    {
-        yield [
-            new UpdateByQuery(
-                new UpdateByQuery\Search(['product'], ['color' => 'olive']),
-                new UpdateByQuery\Update(['color' => 'green']),
-            ),
-            '{"search":{"partial":{"fields":{"color":"olive"}},"types":["product"]},"update":{"fields":{"color":"green"}}}',
-            [
-                'status_url' => '/v1/update_by_query?job_id=1',
-            ],
-            1,
-        ];
-
-        yield [
-            new UpdateByQuery(
-                new UpdateByQuery\Search(['product', 'brand'], ['color' => 'olive']),
-                new UpdateByQuery\Update(['color' => ['green', 'blue'], 'brand' => 'Star']),
-            ),
-            '{"search":{"partial":{"fields":{"color":"olive"}},"types":["product","brand"]},"update":{"fields":{"color":["green","blue"],"brand":"Star"}}}',
-            [
-                'status_url' => '/v1/update_by_query?job_id=12',
-            ],
-            12,
-        ];
-    }
-
     #[Test]
-    #[DataProvider('forUpdateStatus')]
+    #[DataProviderExternal(UpdateByQueryRequestDataProvider::class, 'forUpdateStatus')]
     public function updateStatusPassed(
         int $jobId,
         array $apiResponse,
@@ -91,48 +65,6 @@ class UpdateByQueryRequestTest extends TestCase
         }
 
         $this->assertSame($apiResponse, $response->rawResponse);
-    }
-
-    public static function forUpdateStatus(): iterable
-    {
-        yield [
-            1,
-            [
-                'tracker_id' => 'abcd',
-                'status' => 'complete',
-                'updates_count' => 5,
-                'failures_count' => 0,
-                'failures' => [],
-            ],
-        ];
-
-        yield [
-            111,
-            [
-                'tracker_id' => 'iabad',
-                'status' => 'in progress',
-            ],
-        ];
-
-        yield [
-            12,
-            [
-                'tracker_id' => 'abad',
-                'status' => 'complete',
-                'updates_count' => 5,
-                'failures_count' => 1,
-                'failures' => [
-                    '/products/1' => [
-                        'type' => 'data_schema_mismatch',
-                        'reason' => 'failed to parse [attributes.price]',
-                        'caused_by' => [
-                            'type' => 'number_format_exception',
-                            'reason' => 'For input string: "wrong sale price"',
-                        ],
-                    ],
-                ],
-            ],
-        ];
     }
 
     private function getService(
